[bug/#35]-delete-dup-code & refactoring
