--- src/main/java/yuseteam/mealticketsystemwas/admin/repository/RestaurantRepository.java
@@ -1,7 +1,0 @@
-package yuseteam.mealticketsystemwas.admin.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import yuseteam.mealticketsystemwas.entity.Restaurant;
-
-public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
-}
