+++ src/main/java/yuseteam/mealticketsystemwas/domain/admin/repository/MenuRepository.java
@@ -1,4 +1,4 @@
-package yuseteam.mealticketsystemwas.admin.repository;
+package yuseteam.mealticketsystemwas.domain.admin.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import yuseteam.mealticketsystemwas.entity.Menu;
